CHORE: Move some files from data to config
