+++ src/config/categoriesConfig.ts
