+++ src/config/filter/allCategoriesConfig.ts
