+++ src/config/footerNavigationConfig.ts
